requirements.txt
# Core RAG Stack
google-generativeai>=0.8.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
PyMuPDF>=1.27.2
python-dotenv>=1.2.2

# Evaluation & Science
ragas>=0.1.0
datasets>=2.15.0
pandas>=2.0.0
numpy<2.0.0
rank_bm25>=0.2.2

# UI & Engineering
streamlit>=1.56.0
scikit-learn>=1.8.0
torch>=2.11.0
tqdm>=4.67.0

# Level 4 Observability
arize-phoenix>=14.1.1
openinference-instrumentation-google-genai>=0.1.0
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
